Extentd zidcu interface with executeVoid and executeScalar
This is moving towards allowing easier execution of common query types.
Currently, there is a lot of logic in the code checking the return status of
.execute() and throwing if it's not what we expect.

executeVoid is a wrapper that throws if the result is not SQLITE_DONE
executeScalar is a wrapper that returns null if the result is SQLITE_DONE,
returns the get'd value if it is SQLITE_ROW and throws otherwise.

Some variadic templates have been introduced to make binding arguments easier
and to allow a more generic interface.

A few types have been changed to make binding unambiguous.